Fixed recyclerviewholder variable declarations

+++ app/src/main/java/com/example/android/meallogger/data/PortionDescription.java
@@ -1,6 +1,6 @@
 package com.example.android.meallogger.data;
 
 public class PortionDescription {
-    public int gramWeight;
+    public float gramWeight;
     public String modifier;
 }
